chore: refactor dynamic suggestions prompt and add debug logs
